requirements.txt
flask
jinja2
python-dotenv
waitress
paste